chore: rename test folders to align with player components
